Use movie clip context for storing tracks info
Before this change, we stored our badness scores in the Object context.
Unfortunately that doesn't exist if you don't have any objects, as
pointed out in #2 and #3.

With this change in place, we store our badness scores in the clip
context instead. Now, as long as you have a movie clip loaded, the
plugin's UI will be available.

This should be the proper fix for #2 and #3.